moved discoverInfo into Utils folder
