+++ less/discover.less
@@ -41,11 +41,6 @@
     box-shadow: 10px 10px @metallicGold;
 }
 
-// .card:hover{
-//     box-shadow: 12px 12px @metallicGold;
-// }
-
-
 .infocard{
     display: flex;
     float: right;
